+++ packages/format/src/lib/translations/index.ts
@@ -3,26 +3,26 @@ import { key as key_duration } from './duration/index.js';
 import { key as key_filesize } from './filesize/index.js';
 import { key as key_relativeTime } from './relativeTime/index.js';
 
-export const duration: Translation = {
+export const duration = {
 	key: key_duration,
 	loaders: {
 		en: () => import('./duration/en.js'),
 		fr: () => import('./duration/fr.js')
 	}
-};
+} as const satisfies Translation;
 
-export const filesize: Translation = {
+export const filesize = {
 	key: key_filesize,
 	loaders: {
 		en: () => import('./filesize/en.js'),
 		fr: () => import('./filesize/fr.js')
 	}
-};
+} as const satisfies Translation;
 
-export const relativeTime: Translation = {
+export const relativeTime = {
 	key: key_relativeTime,
 	loaders: {
 		en: () => import('./relativeTime/en.js'),
 		fr: () => import('./relativeTime/fr.js')
 	}
-};
+} as const satisfies Translation;