+++ packages/components/slideshow/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @288-toolkit/slideshow
 
+## 3.2.1
+
+### Patch Changes
+
+- Improve loader types
+
 ## 3.2.0
 
 ### Minor Changes